# Conflicts:
#	RetailersBackSystem-controller/src/main/java/com/guigu/controller/CommodityController.java
#	RetailersBackSystem-service/src/main/java/com/guigu/service/CommodityService.java

+++ RetailersBackSystem-pojo/src/main/java/com/guigu/pojo/Commodity.java
@@ -4,6 +4,7 @@ package com.guigu.pojo;
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -60,7 +61,7 @@ public class Commodity {
     /**
      * 商品库存数量
      */
-    private Double quantity;
+    private Integer quantity;
 
     /**
      * 商品状态 1.在售 2.下架
